// src/module/users/dto/create-user.dto.ts
export interface CreateUserDTO {
  username: string;
  email: string;
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
  role: UsersRole;
  status?: UsersStatus;
}

export interface CreatedUserDTO {
  userId: string;
  username: string;
  email: string;
  firstName: string;
  lastName: string;
  role: UsersRole;
  status: UsersStatus;
  createdAt: Date;
}